% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/API.R
\name{auth}
\alias{auth}
\title{Authentication for Wufoo Name}
\usage{
auth(x)
}
\arguments{
\item{x}{- an empty parameter, e.g. NULL}
}
\description{
method for setting your Wufoo Name permanently
}
\note{
Wufoo currently restricts your API usage to 5000 requests per day.
}
\examples{
options(Wufoo_Name = "johnmalc")
}
\author{
Scott Chamberlain \url{https://github.com/sckott} for his \url{https://github.com/ropensci/rnoaa} package
}

