% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Form.R
\name{form_entries}
\alias{form_entries}
\title{Return Statistics}
\usage{
form_entries(ApiKey, wufoo_name = auth(NULL), formIdentifier = NULL,
  includeTodayCount = "false", showRequestURL = "false",
  systemFields = "true")
}
\arguments{
\item{wufoo_name}{- Default: \code{\link{auth}}}

\item{formIdentifier}{- This placeholder must be replaced with your URL or hash.}

\item{includeTodayCount}{}

\item{showRequestURL}{}

\item{systemFields}{}
}
\value{
EntryId - This value is the unique identifier for your entry.

DateCreated - The date that this entry was submitted.

Created By - The person who created the entry. If submitted through a
form, the value here will be public. If the submission originated in the Entry
Manager this value will be the user name of the submitting user.

DateUpdated - The date that this entry was edited through the Entry
Manager. If the submission has never been updated, this value will be blank.

UpdatedBy - The user name of the person who updated the entry in the
Entry Manager will appear in this element.
}
\description{
If you have 5 submission to your form, you’ll have 5 elements (rows) in the return value from your API call.
}
\examples{
form_entries(ApiKey = "F1QH-Q64B-BSBI-JASJ")
}

