% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Form.R
\name{form_entriesCount}
\alias{form_entriesCount}
\title{Entry Count}
\usage{
form_entriesCount(ApiKey, wufoo_name = auth(NULL), formIdentifier = NULL,
  includeTodayCount = "true", showRequestURL = FALSE)
}
\arguments{
\item{ApiKey}{- User's API key. Go to "https://yourName.wufoo.com/api/code/1/}

\item{wufoo_name}{- Default: \code{\link{auth}}}

\item{formIdentifier}{- this will give you information about just one form.
The call without the "formIdentifier" will return all forms.}

\item{includeTodayCount}{- Will give you today’s entry count for the form.}

\item{showRequestURL}{- use only for DEBUG purposes}
}
\value{
EntryCount - number of entries
}
\description{
Entry Count
}
\examples{
form_entriesCount(ApiKey = "F1QH-Q64B-BSBI-JASJ", formIdentifier = "z5kqx7h1gtvg4g", showRequestURL = TRUE)
}

