% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reports.R
\name{report_entries}
\alias{report_entries}
\title{Returns the entries that make up a specific report.}
\usage{
report_entries(wufoo_name = auth_name(NULL), domain = "wufoo.com",
  systemFields = "true", reportName = NULL, showRequestURL = FALSE,
  debugConnection = 0L)
}
\arguments{
\item{wufoo_name}{- User's Name. Default: \code{\link{auth_name}}}

\item{domain}{- for whatever reason domain's ccTLD may change. With this parameter you can change
default \code{wufoo.com} to another one such as \code{wufoo.eu}}

\item{systemFields}{- return system fields. Default: true}

\item{reportName}{- the name of your report as displayed in the csv export
URL (which is in lowercase with hyphens replacing spaces of your report name). The report should be also public.}

\item{showRequestURL}{- use only for DEBUG purposes}

\item{debugConnection}{- samilar as above. Either "0L" (false; default) or "1L"}
}
\description{
This is essentially an equivalent of the data that would show up in a
datagrid widget on the report, or in an exported copy of the report entry data
}
\examples{
\donttest{
report_entries(reportName = "untitled-report")
}

}
\seealso{
[form_entries()]
}
