% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reports.R
\name{report_entriesCount}
\alias{report_entriesCount}
\title{Returns a count of the entries stored for a specific report.}
\usage{
report_entriesCount(
  wufoo_name = auth_name(NULL),
  domain = "wufoo.com",
  reportName = NULL,
  showRequestURL = FALSE,
  debugConnection = 0L
)
}
\arguments{
\item{wufoo_name}{- User's Name. Default: \code{\link{auth_name}}}

\item{domain}{- for whatever reason domain's ccTLD may change. With this parameter you can change
default \code{wufoo.com} to another one such as \code{wufoo.eu}}

\item{reportName}{- the name of your report as displayed in the csv export
URL (which is in lowercase with hyphens replacing spaces of your report name). The report should be also public.}

\item{showRequestURL}{- use only for DEBUG purposes}

\item{debugConnection}{- samilar as above. Either "0L" (false; default) or "1L"}
}
\description{
This can help with determining the number of elements you have to display.
}
\examples{
\dontrun{
report_entriesCount(reportName = "untitled-report")
}

}
\seealso{
[report_entries()]
}
