% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reports.R
\name{report_fields}
\alias{report_fields}
\title{Returns the field structure for the report's corresponding form.}
\usage{
report_fields(
  wufoo_name = auth_name(NULL),
  domain = "wufoo.com",
  systemFields = "true",
  reportName = NULL,
  showRequestURL = FALSE,
  debugConnection = 0L
)
}
\arguments{
\item{wufoo_name}{- User's Name. Default: \code{\link{auth_name}}}

\item{domain}{- for whatever reason domain's ccTLD may change. With this parameter you can change
default \code{wufoo.com} to another one such as \code{wufoo.eu}}

\item{systemFields}{- return system fields. Default: true}

\item{reportName}{- the name of your report as displayed in the csv export
URL (which is in lowercase with hyphens replacing spaces of your report name). The report should be also public.}

\item{showRequestURL}{- use only for DEBUG purposes}

\item{debugConnection}{- samilar as above. Either "0L" (false; default) or "1L"}
}
\description{
Returns the field structure for the report's corresponding form.
}
\examples{
\dontrun{
report_fields(reportName = "untitled-report")
}

}
\seealso{
[fields_info()]
}
