% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget.R
\name{report_widgets}
\alias{report_widgets}
\title{Returns details of the widgets that make up a specific report.}
\usage{
report_widgets(
  wufoo_name = auth_name(NULL),
  domain = "wufoo.com",
  formIdentifier = NULL,
  showRequestURL = FALSE,
  debugConnection = 0L
)
}
\arguments{
\item{wufoo_name}{- User's Name. Default: \code{\link{auth_name}}}

\item{domain}{- for whatever reason domain's ccTLD may change. With this parameter you can change
default \code{wufoo.com} to another one such as \code{wufoo.eu}}

\item{formIdentifier}{- this will give you information about just one form. A request without
the \code{formIdentifier} will return all forms.}

\item{showRequestURL}{- use only for DEBUG purposes}

\item{debugConnection}{- samilar as above. Either "0L" (false; default) or "1L"}
}
\value{
\url{https://wufoo.github.io/docs/#widgets}
}
\description{
The hash code for a widget can be used to embed the widget using Javascript. Only Chart, Graph,
and Number widgets will be included in the request.
Any Text or Datagrid widgets will not be shown.
}
\examples{
\dontrun{
report_widgets(formIdentifier = "z5kqx7h1gtvg4g")
}

}
