% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Reports.R
\name{reports_info}
\alias{reports_info}
\title{Used to gather details about the reports you have permission to view.}
\usage{
reports_info(ApiKey, wufoo_name = auth(NULL), showRequestURL = FALSE)
}
\arguments{
\item{ApiKey}{- User's API key. Go to "https://yourName.wufoo.com/api/code/1/}

\item{wufoo_name}{- Default: \code{\link{auth}}}

\item{showRequestURL}{- use only for DEBUG purposes}
}
\value{
Name - This is the friendly name you chose when creating this form.

IsPublic - Indicates whether or not the report is accessible through
the Url by the general public. This value is binary (1 = true, 0 = false).

Url - This is the URL for your form. Beware using the URL for API or
linking purposes because it changes with the report title.

Description - Your description of the report.

DateCreated - A timestamp of when the report was created.

DateUpdated - A timestamp of when the report was lasted edited in the Wufoo Report Builder.

Hash - An unchanging hashed value unique to this report on this user's account.
}
\description{
Used to gather details about the reports you have permission to view.
}
\examples{
reports_info(ApiKey = "F1QH-Q64B-BSBI-JASJ", showRequestURL = TRUE)
}

