% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Users.R
\name{user_info}
\alias{user_info}
\title{Return information about the user}
\usage{
user_info(wufoo_name = auth_name(NULL), showRequestURL = FALSE,
  debugConnection = 0L)
}
\arguments{
\item{wufoo_name}{- User's Name. Default: \code{\link{auth_name}}}

\item{showRequestURL}{- use only for DEBUG purposes}

\item{debugConnection}{- either "0L" (false; default) or "1L"}
}
\value{
\url{http://help.wufoo.com/articles/en_US/SurveyMonkeyArticleType/The-Users-API}
}
\description{
Shows information about the user.
}
\examples{
user_info()

}
\references{
wufoo_api - User's API Key. Default: \code{\link{auth_key}}.
}

