% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/API.R
\name{auth}
\alias{auth}
\title{Authentication for Wufoo Name}
\usage{
auth(x)
}
\arguments{
\item{NULL}{for example}
}
\description{
methods written by Scott Chamberlain \url{https://github.com/sckott}
}
\note{
Wufoo currently restricts your API usage to 5000 requests per day.
}

