% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{auth_name}
\alias{auth_key}
\alias{auth_name}
\title{Authentication}
\usage{
auth_name(x)

auth_key(x)
}
\arguments{
\item{x}{- an empty parameter, e.g. NULL}
}
\description{
Method for setting your Wufoo Name and API Key. Your Wufoo name is the subdomain 
of your Wufoo URL displayed in "Account" tab. For example, for http://johnmalc.wufoo.com ->
the Wufoo_Name = "johnmalc". This can be also actually the company's name, e.g. \code{http://google.wufoo.com}

Your API may be found by selecting "Share" for your Form, then 
"API Information" (or go to \code{https://yourName.wufoo.com/api/code/1/}.
}
\note{
Wufoo currently restricts your API usage to 5000 requests per day.
}
\examples{
options(Wufoo_Name = "johnmalc", Wufoo_API = "F1QH-Q64B-BSBI-JASJ")

}
\author{
The code for these methods has been developed by Scott Chamberlain \url{https://github.com/sckott} for his 
\url{https://github.com/ropensci/rnoaa} package. His copyright!
}

