% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{auth_name}
\alias{auth_name}
\alias{auth_key}
\title{Authentication}
\usage{
auth_name(x)

auth_key(x)
}
\arguments{
\item{x}{- an empty parameter, e.g. NULL}
}
\description{
A method for setting your Wufoo Name and API Key. Your Wufoo name is the subdomain
of your Wufoo URL displayed in "Account" tab. For example, for \code{http://johnmalc.wufoo.com},
the \code{Wufoo_Name = "johnmalc"}. This can be the company's name too, e.g. \code{http://google.wufoo.com}

Your API key may be found by selecting "Share" of your form, then
"API Information" (or go to \code{https://yourName.wufoo.com/api/code/1/}.
}
\note{
Wufoo currently restricts free API usage to 100 requests per day.
}
\examples{
options(Wufoo_Name = "johnmalc", Wufoo_API = "F1QH-Q64B-BSBI-JASJ")
}
\author{
The code for these methods has been developed by Scott Chamberlain \url{https://github.com/sckott} for his
\url{https://github.com/ropensci/rnoaa} package.
}
