% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/API.R
\name{auth_name}
\alias{auth_key}
\alias{auth_name}
\title{Authentication}
\usage{
auth_name(x)

auth_key(x)
}
\arguments{
\item{x}{- an empty parameter, e.g. NULL}
}
\description{
Methods for setting your Wufoo Name & API Key, permanently. For both of them visit your
profile. For API Key go to \code{https://yourName.wufoo.com/api/code/1/}
}
\note{
Wufoo currently restricts your API usage to 5000 requests per day.
}
\examples{
options(Wufoo_Name = "johnmalc", Wufoo_API = "F1QH-Q64B-BSBI-JASJ")
}
\author{
The code for these methods has been developed by Scott Chamberlain \url{https://github.com/sckott} for his
\url{https://github.com/ropensci/rnoaa} package. His copyright!
}

