% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Form.R
\name{form_entries}
\alias{form_entries}
\title{Used to gather the data that users have submitted to your form.}
\usage{
form_entries(ApiKey, wufoo_name = auth(NULL), formIdentifier = NULL,
  includeTodayCount = "false", systemFields = "true", sortID = NULL,
  sortDirection = NULL, showRequestURL = FALSE)
}
\arguments{
\item{ApiKey}{- User's API key. Go to "https://yourName.wufoo.com/api/code/1/}

\item{wufoo_name}{- Default: \code{\link{auth}}}

\item{formIdentifier}{- must be replaced with your form's URL or hash.}

\item{includeTodayCount}{- Will give you today’s entry count for the form.}

\item{systemFields}{- return system fields. Default: true}

\item{sortID}{- sort on a single ID, as retrieved from the Fields API.}

\item{sortDirection}{- choose to sort your entries ASC (lowest to highest) or DESC (highest to lowest).}

\item{showRequestURL}{- for debug purposes}
}
\value{
EntryId - This value is the unique identifier for your entry.

DateCreated - The date that this entry was submitted.

Created By - The person who created the entry. If submitted through a
form, the value here will be public. If the submission originated in the Entry
Manager this value will be the user name of the submitting user.

DateUpdated - The date that this entry was edited through the Entry
Manager. If the submission has never been updated, this value will be blank.

UpdatedBy - The user name of the person who updated the entry in the
Entry Manager will appear in this element.
}
\description{
If you have 5 submissions to your form, you’ll have 5 elements (rows) in the return.
}
\examples{
form_entries(ApiKey = "F1QH-Q64B-BSBI-JASJ", formIdentifier = "z5kqx7h1gtvg4g")
form_entries(ApiKey = "F1QH-Q64B-BSBI-JASJ", formIdentifier = "z5kqx7h1gtvg4g", systemFields = "false", showRequestURL = TRUE)
## form_entries(ApiKey = "F1QH-Q64B-BSBI-JASJ", formIdentifier = "z5kqx7h1gtvg4g")
}

