% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Form.R
\name{form_entriesCount}
\alias{form_entriesCount}
\title{Return number of responses to your form}
\usage{
form_entriesCount(wufoo_name = auth_name(NULL), formIdentifier = NULL,
  showRequestURL = FALSE)
}
\arguments{
\item{wufoo_name}{- User's Name. Default: \code{\link{auth_name}}}

\item{formIdentifier}{- this will give you information about just one form. The call without
the "formIdentifier" will return all forms.}

\item{showRequestURL}{- use only for DEBUG purposes}
}
\value{
EntryCount - number of entries
}
\description{
Return number of responses to your form
}
\examples{
form_entriesCount(formIdentifier = "z5kqx7h1gtvg4g", showRequestURL = TRUE)
}
\seealso{
\url{http://help.wufoo.com/articles/en_US/SurveyMonkeyArticleType/The-Entries-GET-API#entrycount}
}

