% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Form.R
\name{form_entriesCount}
\alias{form_entriesCount}
\title{Return number of responses to your form}
\usage{
form_entriesCount(
  wufoo_name = auth_name(NULL),
  formIdentifier = NULL,
  showRequestURL = FALSE,
  debugConnection = 0L,
  domain = "wufoo.com"
)
}
\arguments{
\item{wufoo_name}{- User's Name. Default: \code{\link{auth_name}}}

\item{formIdentifier}{- this will give you information about just one form. A request without
the \code{formIdentifier} will return all forms.}

\item{showRequestURL}{- use only for DEBUG purposes}

\item{debugConnection}{- samilar as above. Either "0L" (false; default) or "1L"}

\item{domain}{- for whatever reason domain's ccTLD may change. With this parameter you can change
default \code{wufoo.com} to another one such as \code{wufoo.eu}}
}
\value{
\url{https://wufoo.github.io/docs/#form-entries-count}
}
\description{
Return number of responses to your form
}
\examples{
\dontrun{
form_entriesCount(formIdentifier = "z5kqx7h1gtvg4g", showRequestURL = TRUE)
}

}
\seealso{
\url{http://help.wufoo.com/articles/en_US/SurveyMonkeyArticleType/The-Entries-GET-API#entrycount}
}
