% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Form.R
\name{form_entriesFromCSV}
\alias{form_entriesFromCSV}
\title{Return number of responses to your form, from CSV format}
\usage{
form_entriesFromCSV(wufoo_name = auth_name(NULL), reportName = NULL,
  showRequestURL = FALSE)
}
\arguments{
\item{wufoo_name}{- User's Name. Default: \code{\link{auth_name}}}

\item{showRequestURL}{- use only for DEBUG purposes}
}
\description{
This function downloads csv file from the url below. the
}
\examples{
form_entriesCount(reportName = "z5kqx7h1gtvg4g", showRequestURL = TRUE)
}
\seealso{
\url{https://YourName.wufoo.com/export/reports/manager/NameOfYourReport.csv}
}

