% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Form.R
\name{form_entriesFromCSV}
\alias{form_entriesFromCSV}
\title{Return responses of your form, from CSV format}
\usage{
form_entriesFromCSV(wufoo_name = auth_name(NULL), reportName = NULL,
  showRequestURL = FALSE)
}
\arguments{
\item{wufoo_name}{- User's Name. Default: \code{\link{auth_name}}}

\item{reportName}{- the name (not the hash!) of the specific report}

\item{showRequestURL}{- use only for DEBUG purposes}
}
\description{
This function downloads csv file from the url below.
The report must be public, without being protected.
E.g. https://YourName.wufoo.com/export/reports/manager/NameOfYourReport.csv
}
\examples{
\dontrun{
options(Wufoo_Name = "johnmalc", Wufoo_API = "F1QH-Q64B-BSBI-JASJ")
df_csv <- form_entriesFromCSV(reportName = "untitled-report", showRequestURL = F)
View(df_csv)
}
}

