% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Form.R
\name{form_info}
\alias{form_info}
\title{Gather details about the forms you have permission to access.}
\usage{
form_info(ApiKey, wufoo_name = auth(NULL), formIdentifier = NULL,
  includeTodayCount = "false", showRequestURL = FALSE)
}
\arguments{
\item{ApiKey}{- User's API key. Go to "https://yourName.wufoo.com/api/code/1/}

\item{wufoo_name}{- Default: \code{\link{auth}}}

\item{formIdentifier}{- this will give you information about just one form. The call without
the "formIdentifier" will return all forms.}

\item{includeTodayCount}{- Will give you today's entry count for the form.}

\item{showRequestURL}{- use only for DEBUG purposes}
}
\value{
\url{http://help.wufoo.com/articles/en_US/SurveyMonkeyArticleType/The-Forms-API}
}
\description{
This API can be used to create a list of all forms belonging to a user and
dynamically generate a form embed snippet to use in your application.
}
\examples{
View(form_info(ApiKey = "F1QH-Q64B-BSBI-JASJ"))
}

