% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Reports.R
\name{reports_info}
\alias{reports_info}
\title{Return details about the reports you have permission to view.}
\usage{
reports_info(wufoo_name = auth_name(NULL), reportName = NULL,
  showRequestURL = FALSE)
}
\arguments{
\item{wufoo_name}{- User's Name. Default: \code{\link{auth_name}}}

\item{reportName}{- the name (not the hash!) of the specific report}

\item{showRequestURL}{- use only for DEBUG purposes}
}
\value{
Name - This is the friendly name you chose when creating this form.

IsPublic - Indicates whether or not the report is accessible through
the Url by the general public. This value is binary (1 = true, 0 = false).

Url - This is the URL for your form. Beware using the URL for API or
linking purposes because it changes with the report title.

Description - Your description of the report.

DateCreated - A timestamp of when the report was created.

DateUpdated - A timestamp of when the report was lasted edited in the Wufoo Report Builder.

Hash - An unchanging hashed value unique to this report on this user's account.
}
\description{
Return details about the reports you have permission to view.
}
\examples{
reports_info(showRequestURL = TRUE)
reports_info(reportName = "untitled-report")
}

