% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Users.R
\name{user_info}
\alias{user_info}
\title{Shows User's Information}
\usage{
user_info(ApiKey, wufoo_name = auth(NULL), showRequestURL = FALSE)
}
\arguments{
\item{ApiKey}{- User's API key. Go to "https://yourName.wufoo.com/api/code/1/}

\item{wufoo_name}{- Default: \code{\link{auth}}}

\item{showRequestURL}{- use only for DEBUG purposes}
}
\value{
\url{http://help.wufoo.com/articles/en_US/SurveyMonkeyArticleType/The-Users-API}
}
\description{
Shows information about the user.
}
\examples{
View(user_info(ApiKey = "F1QH-Q64B-BSBI-JASJ"))
}

