% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Users.R
\name{user_info}
\alias{user_info}
\title{Return information about the user}
\usage{
user_info(wufoo_name = auth_name(NULL), domain = "wufoo.com",
  showRequestURL = FALSE, debugConnection = 0L)
}
\arguments{
\item{wufoo_name}{- User's Name. Default: \code{\link{auth_name}}}

\item{domain}{- for whatever reason domain's ccTLD may change. With this parameter you can change
default \code{wufoo.com} to another one such as \code{wufoo.eu}}

\item{showRequestURL}{- use only for DEBUG purposes}

\item{debugConnection}{- samilar as above. Either "0L" (false; default) or "1L"}
}
\value{
\url{https://wufoo.github.io/docs/#users}
}
\description{
Shows information about the user.
}
\examples{
\donttest{
user_info(debugConnection = 1L, domain = "wufoo.eu")
}

}
\references{
wufoo_api - User's API Key. Default: \code{\link{auth_key}}.
}
